Move BF Suggestion client to separate file
